+++ sphinx/ext/autosummary/__init__.py
@@ -11,7 +11,7 @@ autosummary directive
 The autosummary directive has the form::
 
     .. autosummary::
-       :nosignatures:
+       :no-signatures:
        :toctree: generated/
 
        module.function_1
@@ -237,12 +237,19 @@ class Autosummary(SphinxDirective):
         'caption': directives.unchanged_required,
         'class': directives.class_option,
         'toctree': directives.unchanged,
-        'nosignatures': directives.flag,
+        'no-signatures': directives.flag,
         'recursive': directives.flag,
         'template': directives.unchanged,
+        'nosignatures': directives.flag,
     }
 
     def run(self) -> list[Node]:
+        # Copy the old option name to the new one
+        # xref RemovedInSphinx90Warning
+        # deprecate nosignatures in Sphinx 9.0
+        if 'no-signatures' not in self.options and 'nosignatures' in self.options:
+            self.options['no-signatures'] = self.options['nosignatures']
+
         self.bridge = DocumenterBridge(
             self.env, self.state.document.reporter, Options(), self.lineno, self.state
         )
@@ -462,7 +469,7 @@ class Autosummary(SphinxDirective):
 
         for name, sig, summary, real_name in items:
             qualifier = 'obj'
-            if 'nosignatures' not in self.options:
+            if 'no-signatures' not in self.options:
                 col1 = f':py:{qualifier}:`{name} <{real_name}>`\\ {rst.escape(sig)}'
             else:
                 col1 = f':py:{qualifier}:`{name} <{real_name}>`'